AFMsimulations

Documentation for AFMsimulations.

AFMsimulations.AFM_DMT_experiment

AFMsimulations.AFM_LJ_experiment

AFMsimulations.AFM_eDMT_experiment

AFMsimulations.AFM_vDMT_experiment

AFMsimulations.AFM_vLJ_experiment

AFMsimulations.Cantilever

AFMsimulations.Sample

AFMsimulations.ampl_sweep

AFMsimulations.ampl_sweep

AFMsimulations.eff_young_module

AFMsimulations.f_DMT

AFMsimulations.f_DMT!

AFMsimulations.f_DMT_auto!

AFMsimulations.f_LJ!

AFMsimulations.f_eDMT!

AFMsimulations.f_vDMT!

AFMsimulations.f_vLJ

AFMsimulations.f_vLJ!

AFMsimulations.force_distance

AFMsimulations.force_distance

AFMsimulations.freq_sweep

AFMsimulations.freq_sweep

AFMsimulations.freq_sweep

AFMsimulations.freq_sweep

AFMsimulations.freq_sweep

AFMsimulations.freq_sweep_auto

AFMsimulations.freq_sweep_Φ

AFMsimulations.lock_in_amplifier

AFMsimulations.phase_space

AFMsimulations.savitzky_golay_filter

AFMsimulations.steady_state_check

AFMsimulations.Φ_poincare

AFMsimulations.Φ_poincare_fixpoint

AFMsimulations.AFM_DMT_experiment — Type

DMT experiment, respects elastic behavior within the hertzian contact model

AFMsimulations.AFM_LJ_experiment — Type

Lennard Jones experiment

AFMsimulations.AFM_eDMT_experiment — Type

DMT experiment with exponential damping term –> eDMT

AFMsimulations.AFM_vDMT_experiment — Type

DMT experiment with viscious damping term –> vDMT

AFMsimulations.AFM_vLJ_experiment — Type

Lennard Jones experiment with damping

AFMsimulations.Cantilever — Type

R .. radius of tip, Q .. quality factor of tip, k .. spring constant, E .. Young's module, ν .. poisson coefficient, f_0 .. eigenfrequency.

AFMsimulations.Sample — Type

H .. Hamaker constant, E .. Young's module, ν .. poisson coefficient.

AFMsimulations.ampl_sweep — Method

Amplitude sweeps for DMT-models.

AFMsimulations.ampl_sweep — Method

Amplitude-Sweeps. Keep the frequency constant, instead vary the amplitude of the forcing. This allows us to plot input amplitude against response amplitude. This curve represents a slice in the bifurcation diagramm of the frequncy against response amplitude.

AFMsimulations.eff_young_module — Method

returns effective Young's module of two interacting materials

AFMsimulations.f_DMT! — Method

DMT model. Respects elastic properties via Hertzian contact model.

AFMsimulations.f_DMT — Method

DMT model with StaticArrays.

AFMsimulations.f_DMT_auto! — Method

Autonomous DMT model. Respects elastic properties via Hertzian contact model.

AFMsimulations.f_LJ! — Method

Lennard-Jones model, smooth interaction potential.

AFMsimulations.f_eDMT! — Method

DMT model with additional exponential decaying damping term, inspired by Haviland.

AFMsimulations.f_vDMT! — Method

DMR model with additional viscious damping term, only acting for x > a_0

AFMsimulations.f_vLJ! — Method

Lennard-Jones model, smooth interaction, soft sphere model to avoid singularities at the surface of the sample. Introduces damping via (x - d)^{-3} term. See I. Barke implementation.

AFMsimulations.f_vLJ — Method

Using StaticArrays for faster computations

AFMsimulations.force_distance — Method

returns force-distance relation for DMT experiments

AFMsimulations.force_distance — Method

returns force distance relation for LJ experiments

AFMsimulations.freq_sweep — Method

Forward or backward frequency sweep with constant excitation amplitude for DMT-model.

AFMsimulations.freq_sweep — Method

Forward or backward frequency sweep with constant excitation amplitude for LJ-model.

AFMsimulations.freq_sweep — Method

same for eDMT models

AFMsimulations.freq_sweep — Method

same for vDMT models

AFMsimulations.freq_sweep — Method

Forward or backward frequency sweep with constant excitation amplitude for vLJ-model.

AFMsimulations.freq_sweep_auto — Method

Forward / backward frequency sweep with constant excitation amplitude for DMT-model.

AFMsimulations.freq_sweep_Φ — Method

Frequency sweep of LJ-potential with fixpoints x^{⋆} of Poincare map Φ: x(t) –> x(t + T) with T being the period of the forcing term. Should give even more accurate results and might be more faster (due to less iterations).

AFMsimulations.lock_in_amplifier — Method

Implements digital lock in amplifier. Provide a reference signal (a local osciallator with ω=ωref) and gain information of the phase and amplitude of the input signal. Input:: sig:Array of interest, ωref:reference frequency

AFMsimulations.phase_space — Method

Creates phase space representation for the damped Lennard-Jones model.

AFMsimulations.savitzky_golay_filter — Method

Savitzky-Golay Filter.

AFMsimulations.steady_state_check — Method

Steady state checker, stop simulation if std of amplitude for n periods stays below some tolerance: tol

AFMsimulations.Φ_poincare — Method

Poincare map Φ: x(t) –> x(t + T), with T being the period of the forcing term. For us T = 2π * Ω. Input is a ODE-Problem, initial Vector x, problem specific parameters and Δt.

AFMsimulations.Φ_poincare_fixpoint — Method

Poincare map Φ: x(t) –> x(t + T), with T being the period of the forcing term. For us T = 2π * Ω. Input is a ODE-Problem, initial Vector x, problem specific parameters and Δt. We found a periodic orbit iff |Φ(x) - x|_2 <= ϵ with ϵ being the absolute tolerance. This function can tzhan be used to determine a callback function in an ODEProblem. This way we Should be able to get way more efficient algorithms.