AFMsimulations

Documentation for AFMsimulations.

AFMsimulations.AFM_DMT_experiment

AFMsimulations.AFM_LJ_experiment

AFMsimulations.AFM_eDMT_experiment

AFMsimulations.AFM_vDMT_experiment

AFMsimulations.Cantilever

AFMsimulations.Sample

AFMsimulations.eff_young_module

AFMsimulations.f_DMT!

AFMsimulations.f_LJ!

AFMsimulations.f_eDMT!

AFMsimulations.f_vDMT!

AFMsimulations.freq_sweep

AFMsimulations.freq_sweep

AFMsimulations.freq_sweep

AFMsimulations.freq_sweep

AFMsimulations.AFM_DMT_experiment — Type

DMT experiment, respects elastic behavior within the hertzian contact model

AFMsimulations.AFM_LJ_experiment — Type

Lennard Jones experiment

AFMsimulations.AFM_eDMT_experiment — Type

DMT experiment with exponential damping term –> eDMT

AFMsimulations.AFM_vDMT_experiment — Type

DMT experiment with viscious damping term –> vDMT

AFMsimulations.Cantilever — Type

R .. radius of tip, Q .. quality factor of tip, k .. spring constant, E .. Young's module, ν .. poisson coefficient, f_0 .. eigenfrequency.

AFMsimulations.Sample — Type

H .. Hamaker constant, E .. Young's module, ν .. poisson coefficient.

AFMsimulations.eff_young_module — Method

returns effective Young's module of two interacting materials

AFMsimulations.f_DMT! — Method

DMT model. Respects elastic properties via Hertzian contact model.

AFMsimulations.f_LJ! — Method

Lennard-Jones model, smooth interaction potential

AFMsimulations.f_eDMT! — Method

DMT model with additional exponential decaying damping term, inspired by Haviland.

AFMsimulations.f_vDMT! — Method

DMR model with additional viscious damping term, only acting for x > a_0

AFMsimulations.freq_sweep — Method

Forward or backward frequency sweep with constant excitation amplitude for DMT-model.

AFMsimulations.freq_sweep — Method

Forward or backward frequency sweep with constant excitation amplitude for LJ-model.

AFMsimulations.freq_sweep — Method

same for eDMT models

AFMsimulations.freq_sweep — Method

same for vDMT models